refactored video
